pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notebroom"
version = "0.1.0"
description = "Clean up Jupyter notebook markdown cells using LLMs"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "nbformat>=5.9.2",
    "openai>=1.3.0",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
    "tqdm>=4.66.1",
    "tiktoken>=0.5.1",
]

[project.scripts]
notebroom = "notebroom.notebroom:main"

[tool.hatch.build.targets.wheel]
packages = ["notebroom"]
